update types definition
