+++ src/typings.d.ts
@@ -1,0 +1,34 @@
+declare module "stackjanitor" {
+  export type parameterValue = "Environment";
+  export type EventName = "UpdateStack" | "CreateStack" | "DeleteStack";
+
+  export type parameterKey = {
+    [parameterKey: string]: parameterValue;
+  };
+
+  export interface StackJanitorStatus {
+    [Key: string]: any;
+    results: {
+      stackjanitor: string;
+    };
+  }
+
+  export interface CloudFormationEvent {
+    id: string;
+    detail: {
+      userIdentity: {
+        type: string;
+        sessionContext: {
+          sessionIssuer: {
+            userName: string;
+          };
+        };
+      };
+      eventName: string;
+      requestParameters: {
+        parameters: parameterKey[];
+        stackName: string;
+      };
+    };
+  }
+}
